requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
